chore(origin-ui-core): separate menu and routing
