+++ packages/origin-ui-core/src/components/Account/Account.tsx
@@ -1,91 +1,23 @@
 import React, { useContext } from 'react';
 import { useSelector } from 'react-redux';
-import { NavLink, Route, Redirect } from 'react-router-dom';
+import { Route, Redirect } from 'react-router-dom';
 
 import { PageContent } from '../PageContent/PageContent';
 import { getUserOffchain, getIRecAccount } from '../../features/users/selectors';
-import { AccountSettings } from './AccountSettings';
-import { UserRegister } from './UserRegister';
-import { dataTest, useLinks, useTranslation } from '../../utils';
-import { UserProfile } from './UserProfile';
-import { ConfirmEmail } from './ConfirmEmail';
-import { IRECConnectForm } from '../Organization/IRECConnectForm';
+import { useLinks } from '../../utils';
 import { OriginConfigurationContext } from '..';
-import { OriginFeature } from '@energyweb/utils-general';
+import { accountMenuCreator } from './accountMenuCreator';
 
 export function Account() {
-    const userOffchain = useSelector(getUserOffchain);
-
-    const { baseURL, getAccountLink } = useLinks();
-    const { t } = useTranslation();
-    const { enabledFeatures } = useContext(OriginConfigurationContext);
-
-    const isLoggedIn = Boolean(userOffchain);
-    const organization = useSelector(getUserOffchain)?.organization;
+    const user = useSelector(getUserOffchain);
     const iRecAccount = useSelector(getIRecAccount);
+    const { enabledFeatures } = useContext(OriginConfigurationContext);
+    const { baseURL, getAccountLink } = useLinks();
 
-    const Menu = [
-        {
-            key: 'settings',
-            label: 'settings.navigation.settings',
-            component: AccountSettings,
-            hide: false
-        },
-        {
-            key: 'user-register',
-            label: 'settings.navigation.registerUser',
-            component: UserRegister,
-            hide: isLoggedIn
-        },
-        {
-            key: 'user-profile',
-            label: 'settings.navigation.userProfile',
-            component: UserProfile,
-            hide: !isLoggedIn
-        },
-        {
-            key: 'confirm-email',
-            label: 'settings.navigation.confirmEmail',
-            component: ConfirmEmail,
-            hide: true
-        },
-        {
-            key: 'connect-irec',
-            label: 'settings.navigation.connectIREC',
-            component: IRECConnectForm,
-            hide:
-                !enabledFeatures.includes(OriginFeature.IRec) ||
-                !enabledFeatures.includes(OriginFeature.IRecConnect) ||
-                !organization ||
-                iRecAccount
-        }
-    ];
+    const accountMenuList = accountMenuCreator(user, enabledFeatures, iRecAccount);
 
     return (
         <div className="PageWrapper">
-            <div className="PageNav">
-                <ul className="NavMenu nav">
-                    {Menu.map((menu) => {
-                        if (menu.hide) {
-                            return null;
-                        }
-
-                        return (
-                            <li key={menu.key}>
-                                <NavLink
-                                    exact={true}
-                                    to={`${getAccountLink()}/${menu.key}`}
-                                    activeClassName="active"
-                                    {...dataTest(`account-link-${menu.key}`)}
-                                >
-                                    {t(menu.label)}
-                                </NavLink>
-                            </li>
-                        );
-                    })}
-                </ul>
-            </div>
-
             <Route
                 path={`${getAccountLink()}/:key/:id?`}
                 render={(props) => {
@@ -93,7 +25,7 @@ export function Account() {
 
                     return (
                         <PageContent
-                            menu={Menu.find((item) => item.key === key)}
+                            menu={accountMenuList.find((item) => item.key === key)}
                             redirectPath={getAccountLink()}
                             {...props}
                         />
@@ -104,12 +36,16 @@ export function Account() {
             <Route
                 exact={true}
                 path={`${getAccountLink()}`}
-                render={() => <Redirect to={{ pathname: `${getAccountLink()}/${Menu[0].key}` }} />}
+                render={() => (
+                    <Redirect to={{ pathname: `${getAccountLink()}/${accountMenuList[0].key}` }} />
+                )}
             />
             <Route
                 exact={true}
                 path={`${baseURL}/`}
-                render={() => <Redirect to={{ pathname: `${getAccountLink()}/${Menu[0].key}` }} />}
+                render={() => (
+                    <Redirect to={{ pathname: `${getAccountLink()}/${accountMenuList[0].key}` }} />
+                )}
             />
         </div>
     );
