+++ packages/origin-ui-core/src/components/Account/index.ts
@@ -1,0 +1,3 @@
+export * from './Account';
+export * from './LoginPage';
+export * from './accountMenuCreator';
