+++ packages/origin-ui-core/src/components/Organization/index.ts
@@ -1,0 +1,2 @@
+export * from './Organization';
+export * from './organizationMenuCreator';
