+++ packages/origin-ui-core/src/components/admin/Admin.tsx
@@ -1,52 +1,23 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { NavLink, Redirect, Route } from 'react-router-dom';
+import { Redirect, Route } from 'react-router-dom';
 import { useLinks } from '../../utils';
 import { PageContent } from '../PageContent/PageContent';
-import { AdminUsersTable } from './AdminUsersTable';
-import { AdminUserView } from './AdminUserView';
+import { adminMenuCreator } from './adminMenuCreator';
 
 export function Admin() {
     const { baseURL, getAdminLink } = useLinks();
     const { t } = useTranslation();
 
-    const DevicesMenu = [
-        {
-            key: 'manage-user',
-            label: t('navigation.admin.users'),
-            component: AdminUsersTable
-        },
-        {
-            key: 'user-update',
-            label: 'Update',
-            component: AdminUserView,
-            hide: true
-        }
-    ];
+    const adminMenuList = adminMenuCreator(t);
 
     return (
         <div className="PageWrapper">
-            <div className="PageNav">
-                <ul className="NavMenu nav">
-                    {DevicesMenu.map((menu) => {
-                        if (menu.hide) {
-                            return null;
-                        }
-
-                        return (
-                            <li key={menu.key}>
-                                <NavLink to={`${getAdminLink()}/${menu.key}`}>{menu.label}</NavLink>
-                            </li>
-                        );
-                    })}
-                </ul>
-            </div>
-
             <Route
                 path={`${getAdminLink()}/:key/:id?`}
                 render={(props) => {
                     const key = props.match.params.key;
-                    const matches = DevicesMenu.filter((item) => {
+                    const matches = adminMenuList.filter((item) => {
                         return item.key === key;
                     });
 
@@ -62,14 +33,14 @@ export function Admin() {
                 exact={true}
                 path={getAdminLink()}
                 render={() => (
-                    <Redirect to={{ pathname: `${getAdminLink()}/${DevicesMenu[0].key}` }} />
+                    <Redirect to={{ pathname: `${getAdminLink()}/${adminMenuList[0].key}` }} />
                 )}
             />
             <Route
                 exact={true}
                 path={`${baseURL}/`}
                 render={() => (
-                    <Redirect to={{ pathname: `${getAdminLink()}/${DevicesMenu[0].key}` }} />
+                    <Redirect to={{ pathname: `${getAdminLink()}/${adminMenuList[0].key}` }} />
                 )}
             />
         </div>
