+++ packages/origin-ui-core/src/components/admin/index.ts
@@ -1,0 +1,2 @@
+export * from './Admin';
+export * from './adminMenuCreator';
