+++ packages/origin-ui-core/src/components/certificates/index.ts
@@ -1,0 +1,2 @@
+export * from './Certificates';
+export * from './certificateMenuCreator';
