+++ packages/origin-ui-core/src/components/devices/index.ts
@@ -1,0 +1,3 @@
+export * from './Device';
+export * from './deviceMenuCreator';
+export * from './DeviceSelectors';
