+++ packages/origin-ui-core/src/components/index.ts
@@ -1,23 +1,16 @@
-export * from './Header';
-export * from './certificates/Certificates';
-export * from './devices/Device';
-export * from './Account/Account';
-export * from './ErrorComponent';
-export * from './Modal';
 export * from './PackageConfigurationProvider';
-export * from './admin/Admin';
-export * from './devices/Device';
+export * from './CustomSlider';
+export * from './HierarchicalMultiSelect';
 export * from './ErrorComponent';
-export * from './Modal/RequestCertificatesModal';
-export * from './Modal/AccountMismatchModal';
-export * from './Modal/NoBlockchainAccountModal';
-export * from './Account/LoginPage';
-export * from './Account/Account';
-export * from './Organization/Organization';
-export * from './admin/Admin';
+export * from './IconPopover';
+
+export * from './Account';
+export * from './admin';
+export * from './certificates';
+export * from './devices';
+export * from './Organization';
+
+export * from './Modal';
 export * from './Form';
 export * from './Table';
-export * from './HierarchicalMultiSelect';
-export * from './devices/DeviceSelectors';
-export * from './PageContent/PageContent';
-export * from './CustomSlider';
+export * from './PageContent';
